# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Socle SIRH — Sorbonne-Hôtel",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/reset": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Remettre les données du socle à zéro",
        "description": "Vide le socle, le recharge à l'identique, puis régénère les besoins.\n\nLes schémas des six groupes ne sont pas touchés : le travail des étudiants\nsurvit à une remise à zéro. Les identifiants du socle repartent à 1 et le\njeu de données est reproductible, donc les références logiques stockées par\nles modules (`besoin_id`, `employee_id`…) restent valables.",
        "operationId": "reinitialiser_admin_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reinitialiser Admin Reset Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
    "/admin/besoins/generer": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Rejouer les trois règles de génération des besoins",
        "description": "À lancer après avoir ajusté un seuil dans `rules/seuils.py`.\n\nLes besoins pris en charge ou clôturés par un groupe ne sont jamais\nmodifiés : la commande peut être relancée en plein sprint.",
        "operationId": "generer_besoins_admin_besoins_generer_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generer Besoins Admin Besoins Generer Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
    "/admin/modules/{code}/reset": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Vider et recréer le schéma d'un groupe",
        "description": "Supprime toutes les tables d'un groupe et recrée son schéma, vide.\n\nLe groupe recrée ensuite ses tables depuis la console SQL.",
        "operationId": "reinitialiser_schema_admin_modules__code__reset_post",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reinitialiser Schema Admin Modules  Code  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/etat": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Compter les lignes de chaque table du socle",
        "description": "Le contrôle de cohérence, à passer après chaque remise à zéro.",
        "operationId": "etat_admin_etat_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Etat Admin Etat Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Jeton de groupe": {
        "type": "http",
        "description": "Le jeton remis à votre groupe, en en-tête `Authorization: Bearer <jeton>`.",
        "scheme": "bearer"
      }
    }
  }
}